+++ src/components/rules.js
@@ -1,0 +1,33 @@
+export const rules = {
+    /**
+     * Entity
+     */
+
+    /**
+     * Value Object
+     */
+
+    /**
+     * Domain Service
+     */
+
+    /**
+     * Domain Event
+     */
+
+    /**
+     * Aggregate
+     */
+
+    /**
+     * Repository
+     */
+
+    /**
+     * Factory
+     */
+
+    /**
+     * ACL
+     */
+}
